+++ src/main/java/fr/alasdiablo/diolib/generic/ExperienceRarity.java
@@ -1,0 +1,11 @@
+package fr.alasdiablo.diolib.generic;
+
+/**
+ * Enum use for quantify xp drop by an Entity, Block or Other Element
+ */
+public enum ExperienceRarity {
+    COMMON,
+    UNCOMMON,
+    RARE,
+    NULL
+}