+++ changelog.md
@@ -1,7 +1,15 @@
+### 1.0.4
++ Add more possibility to ConfiguredFeatureHelper
++ Import contributor event from janoeo
++ Change the way of the implementation of IDropExperience interface
+
+### 1.0.3
++ Fix error when using loot table provider
+
 ### 1.0.2
-+ change mod id
-+ change package name
-+ name containing only 6 letters, 3 from the mod name `Diabolo` and 3 from `Lib`
++ Change mod id
++ Change package name
++ Now name containing only 6 letters, 3 from the mod name `Diabolo` and 3 from `Lib`
 
 ### 1.0.1
 + Add loot table provider